Forward ported r3559 [1.2.0] Reverted r3303 (Probably fixed some Linux unicode input junk) as a test for #6735 (Localisation support is broken! tonumber("12.12") returns nil!) r3560 [1.2.0] Reverted r3559 (Reverted r3303 (Probably fixed some Linux unicode input junk) as a test for #6735 (Localisation support is broken! tonumber("12.12") returns nil!)) r3561 [1.2.0] Fixed #6735 (Localisation support is broken! tonumber("12.12") returns nil!) r3562 [1.2.0] Fixed Strafing bug. r3563 [1.2.0] Fixed idle time possibly returning high values just after join. r3564 [1.2.0] Fixed crouch bug Update clients + server. r3565 [1.2.0] Updated credits r3567 [1.2.0] Updated credits #2 r3575 [1.2.0] Added fix/test for a popular crash. Fixed some compile warnings. r3577 [1.2.0] Added fix/test for a popular crash #2 r3578 [1.2.0] Fixed server browser crash and sped up scrolling r3581 [1.2.0] Added fix/test for a popular crash #3 r3582 [1.2.0] Added code to trace timed out messages r3583 [1.2.0] Added fix/test for a popular crash #4 r3584 [1.2.0] Added fix/test for a popular crash #5 r3586 [1.2.0] Changed two data files which were triggering virus false positives r3594 [1.2.0] Set lowest network send priority to medium on the server for a test r3595 [1.2.0] Fixed #6751 (Starting MTA for the first time in some scenarios will cause the client not to receive bikes) r3596 [1.2.0] Updated some debug logging r3597 [1.2.0] Fixed #6754 (Crash when turning threadnet off) r3598 [1.2.0] Added another level of zoom out to the radar map r3599 [1.2.0] Fixed crash introduced in r3595 r3600 [1.2.0] Fixed '*' appearing randomly in GUI r3601 [1.2.0] Added tiny bit more info to crash dumps r3604 [1.2.0] Fixed #6726 (Server with more than 300 players or less disappears from server browser and favourites) r3605 [1.2.0] Re-enabled the ped rotation interpolation in puresync packets. r3606 [1.2.0] Added server bandwidth test code r3609 [1.2.0] Fixed browser search box highlighting

+++ MTA10/core/CServerBrowser.RemoteMasterServer.cpp
@@ -353,7 +353,7 @@ bool CRemoteMasterServer::ParseListVer2 ( CServerListItemList& itemList )
                     uchar ucLength = 0;
                     stream.Read ( ucOffset );
                     stream.Read ( ucLength );
-                    for ( uint i = ucOffset ; i  < ucOffset + ucLength ; i++ )
+                    for ( uint i = ucOffset ; i < (uint)( ucOffset + ucLength ) ; i++ )
                         if ( i < pItem->strSearchableName.length () )
                             pItem->strSearchableName[i] = '\1';
                 }
